+++ src/Workspaces/Core/Portable/FindSymbols/Declarations/DeclarationFinder.cs
@@ -64,8 +64,8 @@ namespace Microsoft.CodeAnalysis.FindSymbols
                 // exact name search, then we will run the query's predicate over every DeclaredSymbolInfo stored in
                 // the doc.
                 var containsSymbol = isExactNameSearch
-                    ? await ContainsSymbolsWithNameAsync(project, query.Name, cancellationToken).ConfigureAwait(false)
-                    : await ContainsSymbolsWithNameAsync(project, query.GetPredicate(), filter, cancellationToken).ConfigureAwait(false);
+                    ? await project.ContainsSymbolsWithNameAsync(query.Name, cancellationToken).ConfigureAwait(false)
+                    : await project.ContainsSymbolsWithNameAsync(query.GetPredicate(), filter, cancellationToken).ConfigureAwait(false);
 
                 if (!containsSymbol)
                     return;
@@ -89,76 +89,6 @@ namespace Microsoft.CodeAnalysis.FindSymbols
 
                 list.AddRange(FilterByCriteria(symbolsWithName, filter));
             }
-        }
-
-        private static Task<bool> ContainsSymbolsWithNameAsync(
-            Project project, string name, CancellationToken cancellationToken)
-        {
-            return ContainsSymbolsAsync(
-                project,
-                (index, cancellationToken) => index.ProbablyContainsIdentifier(name) || index.ProbablyContainsEscapedIdentifier(name),
-                cancellationToken);
-        }
-
-        private static Task<bool> ContainsSymbolsWithNameAsync(
-            Project project, Func<string, bool> predicate, SymbolFilter filter, CancellationToken cancellationToken)
-        {
-            return ContainsSymbolsAsync(
-                project,
-                (index, cancellationToken) =>
-                {
-                    foreach (var info in index.DeclaredSymbolInfos)
-                    {
-                        if (FilterMatches(info, filter) && predicate(info.Name))
-                            return true;
-                    }
-
-                    return false;
-                },
-                cancellationToken);
-
-            static bool FilterMatches(DeclaredSymbolInfo info, SymbolFilter filter)
-            {
-                switch (info.Kind)
-                {
-                    case DeclaredSymbolInfoKind.Namespace:
-                        return (filter & SymbolFilter.Namespace) != 0;
-                    case DeclaredSymbolInfoKind.Class:
-                    case DeclaredSymbolInfoKind.Delegate:
-                    case DeclaredSymbolInfoKind.Enum:
-                    case DeclaredSymbolInfoKind.Interface:
-                    case DeclaredSymbolInfoKind.Module:
-                    case DeclaredSymbolInfoKind.Record:
-                    case DeclaredSymbolInfoKind.RecordStruct:
-                    case DeclaredSymbolInfoKind.Struct:
-                        return (filter & SymbolFilter.Type) != 0;
-                    case DeclaredSymbolInfoKind.Constant:
-                    case DeclaredSymbolInfoKind.Constructor:
-                    case DeclaredSymbolInfoKind.EnumMember:
-                    case DeclaredSymbolInfoKind.Event:
-                    case DeclaredSymbolInfoKind.ExtensionMethod:
-                    case DeclaredSymbolInfoKind.Field:
-                    case DeclaredSymbolInfoKind.Indexer:
-                    case DeclaredSymbolInfoKind.Method:
-                    case DeclaredSymbolInfoKind.Property:
-                        return (filter & SymbolFilter.Member) != 0;
-                    default:
-                        throw ExceptionUtilities.UnexpectedValue(info.Kind);
-                }
-            }
-        }
-
-        private static async Task<bool> ContainsSymbolsAsync(
-            Project project, Func<SyntaxTreeIndex, CancellationToken, bool> predicate, CancellationToken cancellationToken)
-        {
-            var tasks = project.Documents.Select(async d =>
-            {
-                var index = await SyntaxTreeIndex.GetIndexAsync(d, cancellationToken).ConfigureAwait(false);
-                return index != null && predicate(index, cancellationToken);
-            });
-
-            var results = await Task.WhenAll(tasks).ConfigureAwait(false);
-            return results.Any(b => b);
         }
 
         private static async Task AddMetadataDeclarationsWithNormalQueryAsync(
